+++ src/main/java/aventura/app/exceptions/MiEntrdadaSalidaException.java
@@ -1,0 +1,7 @@
+package aventura.app.exceptions;
+
+public class MiEntrdadaSalidaException extends Exception {
+    public MiEntrdadaSalidaException(String message) {
+        super(message);
+    }
+}
